Hyperfix on locking

+++ cluster/config.go
@@ -147,14 +147,14 @@ func (c *ContainerConfig) Constraints() []string {
 	return c.extractExprs("constraints")
 }
 
-func (c *ContainerConfig) CpusUsed() []int64 {
-	cores := make([]int64, 0)
+func (c *ContainerConfig) CpusUsed() []int {
+	cores := make([]int, 0)
 	for sCore := range strings.Split(c.Cpuset, ",") {
-		cores = append(cores, int64(sCore))
+		cores = append(cores, sCore)
 	}
 	return cores
 }
 
-func (c *ContainerConfig) NCpus() int64 {
-	return int64(len(c.CpusUsed()))
+func (c *ContainerConfig) NCpus() int {
+	return len(c.CpusUsed())
 }
